Add some more pages and js/css tweaks
Add an "all posts" page that will link to all the different posts you
can see.

Add first post! That had some images, so add styling for figures and
figcaptions and the actual images.

Sort out the home page so that it's not lorem ipsum, but actually says
something.

Remove scrolling on the page when the big menu is open.
Before you could still scroll and it was annoying.